模板功能差一点

+++ app/src/main/java/com/weilai/jigsawpuzzle/weight/template/TemplateViewInfo.java
@@ -7,10 +7,10 @@ import com.luck.picture.lib.entity.LocalMedia;
 import com.weilai.jigsawpuzzle.keep.IBean;
 
 public class TemplateViewInfo implements IBean {
-    private  int mLeftX;
-    private  int mLeftY;
-    private  int mRightX;
-    private  int mRightY;
+    private  float mLeftX;
+    private  float mLeftY;
+    private  float mRightX;
+    private  float mRightY;
     private  int mPosition;
     private  boolean mHasPic;
     private  String mUrl;
@@ -21,7 +21,7 @@ public class TemplateViewInfo implements IBean {
     public void setLocalMedia(LocalMedia localMedia) {
         this.localMedia = localMedia;
     }
-    public TemplateViewInfo(int leftX,int leftY,int rightX,int rightY,int position,boolean hasPic,String url){
+    public TemplateViewInfo(float leftX,float leftY,float rightX,float rightY,int position,boolean hasPic,String url){
         this.mHasPic = hasPic;
         this.mLeftX = leftX;
         this.mPosition = position;
@@ -61,19 +61,19 @@ public class TemplateViewInfo implements IBean {
         this.mUrl = mUrl;
     }
 
-    public int getLeftX() {
+    public float getLeftX() {
         return mLeftX;
     }
 
-    public int getLeftY() {
+    public float getLeftY() {
         return mLeftY;
     }
 
-    public int getRightX() {
+    public float getRightX() {
         return mRightX;
     }
 
-    public int getRightY() {
+    public float getRightY() {
         return mRightY;
     }
